% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optims.r
\name{optims}
\alias{optims}
\title{Try different optimization procedures}
\usage{
optims(par, fn = NULL, ineq.fn = NULL, eq.fn = NULL, target = NULL,
  lower = NULL, upper = NULL, gr = NULL, algos = selector(par = par, fn
  = fn, gr = gr, ineq.fn = ineq.fn, eq.fn = eq.fn, lower = lower, upper =
  upper), selector = select.base.algos, tol = 1e-08, constr.tol = 1e-12,
  try.all.algos = !TRUE, store.summary = try.all.algos, ...)
}
\description{
Try different optimization procedures
}

